+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Improved canonicalization performance
+- Reduced attachment memory allocation
 
 ## [2.1.1] 2020-11-16
 ### Changed