+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
 ### Added
 - `helper.FreeOSMemory()` to explicitly call the GC and release the memory to the OS
 
 ### Changed
+- users of the library no longer need a `replace` directive for x/crypto
 - added new calls to `runtime.GC()` in the low memory attachment processor
 
 ## [2.1.2] 2020-12-01