+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Changed
+- Removed an unecessary cloning in the attachment processor, to perform better in low memory settings
+
 ## [2.1.4] 2021-01-08
 ### Added
 - Methods for generating an verifying encrypted detached signatures