+++ CHANGELOG.md
@@ -4,13 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.1.6] 2021-03-17
 ### Added
 - Decryption tests for attachments
 
 ### Fixed
 - Armoring headers for public or private keys
 - Session key decoding on invalid keys
+- Support for multiple embedded signatures (not nested)
+
+### Security
+- Updated the underlying crypto library
 
 ## [2.1.5] 2021-02-19